+++ assets/src/ba_data/python/bastd/ui/store/item.py
@@ -92,6 +92,10 @@ def instantiate_store_item_display(
         tint_tex = character.icon_mask_texture
         title_v = 0.255
         price_v = 0.145
+    elif item_name == 'merch':
+        base_text_scale = 0.6
+        title_v = 0.85
+        price_v = 0.15
     elif item_name in ['upgrades.pro', 'pro']:
         base_text_scale = 0.6
         title_v = 0.85
@@ -165,6 +169,23 @@ def instantiate_store_item_display(
             tint2_color=tint2_color,
         )
 
+    if item_name == 'merch':
+        frame_size = b_width * 0.7
+        im_dim = frame_size * (100.0 / 113.0)
+        im_pos = (
+            b_pos[0] + b_width * 0.5 - im_dim * 0.5 + b_offs_x,
+            b_pos[1] + b_height * 0.5 - im_dim * 0.5,
+        )
+        ba.imagewidget(
+            parent=parent_widget,
+            position=im_pos,
+            size=(im_dim, im_dim),
+            transition_delay=delay,
+            draw_controller=btn,
+            opacity=1.0,
+            texture=ba.gettexture('merch'),
+        )
+
     if item_name in ['pro', 'upgrades.pro']:
         frame_size = b_width * 0.5
         im_dim = frame_size * (100.0 / 113.0)
@@ -184,7 +205,6 @@ def instantiate_store_item_display(
         )
         txt = ba.Lstr(resource='store.bombSquadProNewDescriptionText')
 
-        # t = 'foo\nfoo\nfoo\nfoo\nfoo\nfoo'
         item['descriptionText'] = ba.textwidget(
             parent=parent_widget,
             text=txt,