+++ src/ballistica/ballistica.cc
@@ -32,7 +32,7 @@
 namespace ballistica {
 
 // These are set automatically via script; don't modify them here.
-const int kAppBuildNumber = 20966;
+const int kAppBuildNumber = 20967;
 const char* kAppVersion = "1.7.16";
 
 // Our standalone globals.